+++ .php-cs-fixer.dist.php
@@ -1,13 +1,9 @@
 <?php
 
-$finder = (new PhpCsFixer\Finder())
-    ->in([
-        __DIR__ . '/src',
-        __DIR__ . '/tests',
-    ]);
+$finder = new PhpCsFixer\Finder()->in([__DIR__ . "/src", __DIR__ . "/tests"]);
 
-return (new PhpCsFixer\Config())
+return new PhpCsFixer\Config()
     ->setRules([
-        '@PER-CS' => true,
+        "@PER-CS" => true,
     ])
     ->setFinder($finder);